Move transform logic into its own submodule
