+++ tsdat/transform/adi.py
